Fix colors rarely changing in almost random ways
Ever since color pair compression was added in commit
b396e661f85b36e9cd4664fe7b6da3e1f7fb2e0d, it was wrong.

Behaviour in tests and in the main code differed in one crucial way:
.move_pair() in tests simulated actually moving pair value while the
code assumed that the move has already been done and just take this
information into account.

This caused some color pairs to take on almost random value, but because
most remained as they should and also because this happens rarely, this
went unnoticed.

Recently, commit 1e413bcc2b736655523adfedc567f869a9c38361 was made
because this actually become visible when using chafa and such, but it
wasn't understood to be a sign of a bug.  While the commit was working
around this behaviour, it actually made it more apparent because now
every full redraw had a chance of triggering it (although :restart was
the most reliable way).

Actually move the pair before calling .move_pair().