Handle relative paths better in paths_are_equal()
"a" and "./a" were canonicalized to themselves.  Fixing this in
canonicalize_path() to always add "./" breaks its uses for things like
patterns in auto-commands.  Always stripping leading "./" can
potentially result in running a command with "-path" instead of
"./-path".  Although not sure the latter is possible, decided not to
risk.