Fix handling mix of links and spaces
Previously, "fnamemodify()" and "resolve()" were run after
"fnameescape()". This does not work with escaped characters - these
functions don't know how to handle escaped filenames.

In other cases, "fnamemodify()" and "resolve()" were used correctly, but
"fnameescape()" was never called, causing it to break with filenames
with escaped characters.

In other cases, "resolve()" was never used. For these cases, the outcome
was the same, but this allows all file names to be used as buffer names,
which is required for an upcoming fix.