Warn user about incorrect string check in conds
In commands like :if or :elseif, lone `$VAR` or `"bla"` are converted to
bool by checking the string for being empty instead of by converting it
to an integer.

The bug will be fixed after making the release to let users a chance to
update their configurations.

Thanks to justpretending2.

Closes #1023 on GitHub.