Don't set NULL description in bg_run_external_job
Nothing happens in this case as is, but it's more of an accident because
job->descr is NULL initially and replace_string() skips replacing a
value with itself.  Do not rely on this "optimization" which might
change in the future.  This is mostly a concern for Lua API where
corresponding flag can be provided without the description.

Coverity pointed to this, although made a mistake of thinking there is
an actual issue.